googlecode issue 22 Cyclic references will cause a stack overflow
The mutable type to field checker now behaves differently if there is a circular reference between classes currently being analysed. It will report the assignment as causing the class to be mutable. 

Also fixed a bug where the descriptors of arrays of reference types would not load the class correctly.



git-svn-id: http://mutability-detector.googlecode.com/svn@499 c7a0535c-eda6-11de-83d8-6d5adf01d787