Big checkin. Whoops.
Change MutableTypeToFieldChecker to handle case where PUTFIELDs are performed on classes other than the one being analysed. Do this by accepting a raw signature, rather than failing because we haven't seen the real type signature. Bit of a hack, but there you go.

Also add the option to supply a '-failFast' flag to the command line interface, to toggle whether to let exceptions propagate immediately, or just have results be an error kind.

git-svn-id: http://mutability-detector.googlecode.com/svn@669 c7a0535c-eda6-11de-83d8-6d5adf01d787