Fix for GitHub issue #38 - unit tests do not work with sbt.
The "sbt test" task uses a custom classloader when running the unit test. Mutability Detector, when it couldn't load a class as a stream, tried to load it from the classloader, via asm's ClassReader constructor, which uses the system classloader. Fix is to obtain the stream by grabbing the same classloader that has loaded Mutability Detector. This should be fairly safe.