First attempt at making MutabilityAssert.assertX methods threadsafe.
Instead of tracking which analysis is in progress by adding and removing from a list, introduce the 'AnalysisInProgress' type to contain this. It is itself immutable, and is only modified when making a request for a transitive analysis. Checkers will be given an immutable AnalysisInProgress, and pass it along when requesting MutableTypeInformation. This means multiple threads performing multiple analyses will not share the state of what is in progress within a single MutableTypeInformation instance.