Consider immutable classes for generic fields
Given a class with a generic type parameter T, where T is guaranted to
be an instance of U. If the class has a final field of type T and U has
been guaranteed to be immutable, then the class will also be considered
immutable.

This implementation modifies the "allowing" matcher only. It uses
reflection to determine if the field type being assigned had previously
been specified immutable.

Addresses: #104