Add support for Java 16
This upgrades the ASM and ByteBuddy (via Mockito) dependencies so they are
compatible with Java 16. To test, I built this library with the two current
LTS Java versions (8 and 11) as well as Java 16, the latest version. In
addition, I tested by running the `ClientOfMutabilityDetector` tests with
Java 16 while pointing to a local version of `MutabilityDetector` built with
Java 8.

Resolves: #181